+++ plugins/Feedback/tests/Integration/FeedbackTest.php
@@ -39,7 +39,6 @@ class FeedbackTest extends IntegrationTestCase
             'a98732d98732',
             'user1@example.com',
             'user1',
-            'ab9879dc23876f19',
             '2019-03-03'
         );
         FakeAccess::$identity = 'user1';