+++ plugins/Overlay/javascripts/Overlay_Helper.js
@@ -29,7 +29,7 @@ var Overlay_Helper = {
 
         var token_auth = piwik.broadcast.getValueFromUrl("token_auth");
         if (token_auth.length && piwik.shouldPropagateTokenAuth) {
-            url += '&token_auth='  + encodeURIComponent(token_auth);
+            url += '&force_api_session=1&token_auth='  + encodeURIComponent(token_auth);
         }
 
         if (link) {