+++ plugins/UserCountry/Columns/Country.php
@@ -82,8 +82,8 @@ class Country extends Base
     public function onNewVisit(Request $request, Visitor $visitor, $action)
     {
         $value = $this->getUrlOverrideValueIfAllowed('country', $request);
-
         if ($value !== false) {
+            $value = substr($value, 0, 3);
             return $value;
         }