+++ tests/PHPUnit/System/AllWebsitesTest.php
@@ -35,7 +35,7 @@ class AllWebsitesTest extends SystemTestCase
             UsersManagerAPI::getInstance()->addUser('limitedUser', 'smartypants', 'user@limited.com');
             UsersManagerAPI::getInstance()->setUserAccess('limitedUser', 'view', array(2, 3));
             $userModel = new UsersManagerModel();
-            $userModel->updateUserTokenAuth('limitedUser', self::$userTokenAuth);
+            $userModel->addTokenAuth('limitedUser', self::$userTokenAuth, 'desc', '2020-01-02 03:04:05');
         }
     }